test(icom): re-home power derivation coverage Principle XI.
